+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- Groups - Prompt Login button for unverified users [#357](https://github.com/rokwire/illinois-app/issues/357).
 - Groups Search [#371](https://github.com/rokwire/illinois-app/issues/371).
 ### Changed
 - Changes to Create Group Event [#345](https://github.com/rokwire/illinois-app/issues/345).